Commit # 9
Amended the input text file to include two arguments.
Updated the output text file to include the text output.

Found a bug that kept the code from reading EOF
(it was sneaky too, debugger kept skipping over it)

Now the project is completely done and ready to turn in a day early for an extra point!